[Attributor][FIX] `byval` arguments of `readnone` functions are read
At the call site of a function that is `readnone` we can still read
memory if one of the arguments is passed `byval` because a copy is made.
While we had tests for this, we did not check for the right outcome.
This fix is slightly involved as we cannot only look at the result of
`Instruction::mayReadOrWriteMemory` when we determine instructions that
may read or write memory. In a follow up we will try to move the logic
into `mayReadOrWriteMemory` to avoid this "hack".