[OpenMP] Provide a flag to disable safety checks for GPU optimizations
The optimization introduced in D83271 will verify a safety and a
precautionary condition. Until we find a proper solution for the former
we allow users to disable the check. If they do it and violate the
problematic (but probably unlikely) case, they will see UB, most likely
a crash when a non-function pointer is called. The flag has the word
`unsafe` in the name and the remarks are adjusted to provide some
insight what is happening.

Differential Revision: https://reviews.llvm.org/D83832