[Attributor][FIX] Do not look for users of "non-locals" in the MBEC
When we walk the must-be-executed-context (MBEC) we try to skip
uninteresting values by only looking at users of the associated value.
This only makes sense, actually the entire MBEC idea does, if the value
is "local". We cannot derive information about a global or constant from
some uses of it.