[CallGraphUpdater][FIX] Do not attempt to update unreachable functions
In the sqlite3 build, under certain conditions, we ended up with an CG
update order in which an internal function ended up being unreachable
before `CallGraphUpdate::reanalyzeFunction` was called on it. That
resulted in a segfault as the function did not belong to any SCC
anymore. We now detect this and avoid to reanalyze such unreachable
functions.

I failed to simplify the 1.9MB sqlite3.bc file as conversion into .ll,
e.g., as performed by llvm-reduce, causes the problem to disappear.