Diagnose an unavailable `long double` type properly
This is pretty much the same as the __float128 handling. To avoid noise
we will not emit an error for CUDA `long double` types now. The NVPTX
target states `nvcc` warns and treats `long double` as `double`, we
could warn were the TODO was added. The TODO is taken from the CUDA
special case for float16.