[OpenMP] Unify omptarget level APIs wrt. the `AsyncInfoTy` usage
This patch unifies our libomptarget API in two ways:

- Always pass a AsyncInfoTy object, the Queue member decides if it is in
  use or not, thus if there are outstanding actions.
- Almost always synchronize in the interface layer and not in the
  omptarget layer. There is a TODO at the point we can't do this yet.

A side effect is that we now put all constructor and static initializer
kernels in a stream too, if the device utilizes them via the AsyncInfoTy.

Site note: On a V100 system the GridMini performance for small sizes
more than doubled.

Differential Revision: https://reviews.llvm.org/D96379