[Attributor][FIX] Do not use at uses outside the SCC in CGSCC runs
It is not 100% clear if we can look at uses outside the SCC but even if
we have to be careful not to assume that we could later change those
uses. This fixes a bug in the existing tests where we thought we would
replace a call with a constant and therefore could drop the constant in
the return values. However, the call was outside the SCC and we did not
modify it, causing a mismatch = bug. Giving up on uses outside the SCC
might be a little conservative but OK for now.