[Utils] Check for generated functions inline if possible
With D83004 we allowed to check for generated functions. However, since
those might be generated in various places we did simply check for all
functions in the end. Furthermore, this prevented us from reusing check
prefixes across run lines. For every run line we checked for all functions
regardless if the functions were equal across prefixes.

To reduce the number of check lines we now allow to reuse them across
run lines. That is, we can have prefixes that span multiple run lines
even when we check for generated functions. This will only reduce the
run lines if we can find a total order of the generated functions
though. The check lines for generated functions are placed in-between
check lines of existing functions or at the end, as necessary.

The IR update script needs adjusts similar to the CC one.

Differential Revision: https://reviews.llvm.org/D102558