[Attributor] Allow to materialize a simplified value
We already allow AAValueSimplify to look through instructions and such.
Now we also enable it to create new instructions in order to manifest
simplification results. We basically copy whatever is necessary into a
new context and rewrite it.

This will become more interesting once we track a local and a remote
simplified value.