[OpenMP] Avoid starting a kernel without work
For target regions that have a combined `distribute` we emit a
`__kmpc_push_target_tripcount_mapper` call prior to the
`__tgt_target_...` call. While we need to always honor the `map`
clauses we can avoid starting an actual kernel if the loop trip count
associated with the distribute is 0. This is not an optimization for
the 0 trip count case but rather the reason we can emit code in the
target that does not need worry about 0 trip count distribute loops.