[OpenMP][NFCI] Record the location string as part of the ident_t
When we want to inspect ident_t objects from the device on the host we
need to copy the string over. So far we did not know how long the string
was which prevented an efficient way of copying it. Given that we have
two unused fields in the struct we can just store the string size.

NOTE: Tests will be adjusted shortly.

Differential Revision: https://reviews.llvm.org/D112493