[OpenMP][1/3] Introduce the KernelEnvironment into the device runtimes
This introduces the kernel environment which contains information passed
by the compiler to a GPU kernel. For now it mostly encapsulated the
ident_t object and the execution configuration, thus information we
passed explicitly before. We will add more content, including mutable
content similar to the debug indention level, later on.

Differential Revision: https://reviews.llvm.org/D116908