[OpenMP][2/3] Introduce the KernelEnvironment into the middle-end
This introduces the kernel environment, which contains information passed
by the compiler to a GPU kernel, to the middle-end (=LLVM/Core). This
allows us to create the environment and change (especially the execution
configuration part) during OpenMP-Opt. OpenMP-IR-Builder changes are
included to manifest the new globals and create the initial ones in the
first place.

Differential Revision: https://reviews.llvm.org/D116909