[MemoryBuiltins][FIX] Adjust index type size properly wrt. AS casts
We strip AS casts during the computation of size and offset in the
ObjectSizeOffsetVisitor. This is fine but we need to adjust the APInt
properly to avoid mismatches. This code ensures the caller of compute
sees APInts that match the index type size of the value passed to
compute, not the value result of the strip pointer cast.

NOTE: Ideas for tests are appreciated. My reproducer looks something
      like this: DSE runs on some stores to a pointer which is a
      select of two AS casts, on one side a GEP on the other an alloca.
      Unsure if/how I can trigger DSE with minimal example while causing
      it to run getObjSize (though I have not tried).

Differential Revision: https://reviews.llvm.org/D118727