[OpenMP] Use assumed underlying objects for SPMD-zation
We used the non-optimistic `getUnderlyingObjects` function before which
does not work inter-procedurally. However, we can easily get into
situations that would require such a reasoning for SPMD-zation. One
example is a complex reduction in a generic target region since the
complex constructor is called to initialize some thread local memory.
Since thread local accesses are SPMD-compatible we can convert such a
kernel but only if we look at the underlying objects passed to the
constructor. Luckily, `AA::getAssumedUnderlyingObjects` does all that.