[OpenMP] Allow SPMD-zation for guard-free functions with complex uses
If we need guards for instructions we have to ensure our guarding scheme
is able to deal with all contexts. To this end we are very restrictive
and only allow the same kind of execution mode (SPMD or generic) and
no use inside of parallel regions. However, all of this is only
necessary if we actually perform guarding. If there is no instruction to
guard we can allow the code to be called from different contexts without
any issue.