[OpenMP][WIP] Explicitly deinitialize device resources
This is a WIP patch, some parts are missing and I need feedback on it
first.

There are two problems this patch tries to address:
1) We currently free resources in a random order wrt. plugin and
   libomptarget destruction. This patch should ensure the plugins can
   always relinquish their resources first as we call into them during
   the libomptarget DeviceTy destructor.
2) We need to support (hard) pause runtime calls eventually. This patch
   allows us to free all associated resources, though we cannot
   reinitialize the device yet.

The only known weirdness is the event pool which is shared across
devices. We could have one per device, or, deallocate it when the last
device associated with the plugin is deallocated. I opted for the latter
now.

Differential Revision: https://reviews.llvm.org/D120089