[Attributor] Avoid replacing return operands twice
As replacements will become more complex it is better to have a single
AA responsible for replacing a use. Before this patch
AAValueSimplify* and AAValueSimplifyReturned could both try to replace
the returned value. The latter was marginally better for the old pass
manager and when a function was already carrying a `returned` attribute
and when the context of the return instruction was important. The latter
will be resolved shortly, the former is an unrealistic use case to begin
with.