[IRMover] Do not drop names of struct types that are used
When we link bitcode via clang (`-mlink-bitcode-file`) we can end up
removing the name of a struct type that is present in the user code
and the linked in file. This occurred for OpenMP codes but turned out
to be really hard to reproduce. The necessary code change to avoid
dropping those names is simple though and hopefully uncontroversial.

Differential Revision: https://reviews.llvm.org/D122328