[OpenMP] Only initialize a single queue/stream/event eagerly
Initialization for many queues/streams/events might come at a cost even
if we do not use them. This patch lazily initializes them, otherwise
nothing (major) is supposed to change. A minor difference is the
handling of an error in the initialization of an AMD queue (other than
the first). We now report an error but continue with the first queue;
unsure if this will ever come up.