[Attributor] Replace AAReturnedValues with AAPotentialValuesReturned
The very first AA, at least the first one in order, is now gone.
AAReturnedValues was from a different time, one might say, a simpler
time. However, it was already rewamped once to use
`Attribute::getAssumedSimplifiedValues`, which is what the updated
AAPotentialValuesReturned now does too. To match the old behavior we
needed to avoid the helper AAReturnedFromReturnedValues and iterate the
return instructions explicitly, however, it is still less complexity
than it was before.

Fixes: https://github.com/llvm/llvm-project/issues/63404