[OpenMP] Associate the KernelEnvironment with the GenericKernelTy
By associating the kernel environment with the generic kernel we can
access middle-end information easily, including the launch bounds ranges
that are acceptable. By constraining the number of threads accordingly,
we now obey the user provided bounds that were passed via attributes.
The next step is to deal with the extra warp that Generic mode
"requires".