GPU Sanitizer support for global variables
This commit extends GPU sanitizer support to global variables. Shadow variables that store the fake pointer values are added for each user-defined global. Right now, globals are registered by the host and de-allocated by the device, but this is expected to change in the future. The device-side allocation function is in the instrumentation pass, but it's commented out for now. Global variable references are rewritten as load instructions that get their respective fake pointer from the shadow globals.