[Attributor] Introduce AALoadsFromStore and AAStoresFromLoad
These are caching wrappers around the former
`AA::getPotentialCopiesOfStoredValue` and
`AA::getPotentiallyLoadedValues` in an effort to avoid costly
recomputation. There is no functional change intended as users can
interact with the AAs the same way they used the functions before.

The test change is due to a timeout (32 iterations) which we need to
address later.