[AMDGPU][Attributor] Fix AAAMDWavesPerEU deduction
This should replaces #114357.

The problem this resolves is that AAAMDWavesPerEU used "assumed"
information from AAAMDFlatWorkGroupSize to derive "known" information.
This is generally not valid. The cannonical way is to use
"assumed"/"known" information to improve "assumed"/"known" information,
respectively. What happend before was that the new "known" information
(derived from "assumed") was manifested even though it should not have
been. That is how we ended up with external functions that had
"amdgpu-waves-per-eu" set. This won't happen if we only set the assumed
information as the invalidation, caused by unknown callers, will
fallback to the "unset" known information.

In addition to the above, we will identify invalid waves-per-eu values,
emit a warning, and replace the invalid values with valid ones. This
happens also for kernels. See
`llvm/test/CodeGen/AMDGPU/propagate-waves-per-eu.ll`

This should hopefully fix #123092.