[Attributor][NFC] Make offsets in AAPointerInfo ranges
We used to track offsets as simple integers, which is fine except if we
want to deal with ranges, i.a., accesses in loops. We now track offsets
as ranges without ever using a range of size larger than 0. This also
lead to restructuring the handling of offsets down the line. We keep
them separate longer and do not merge in the access size as it is
different, conceptually, from the pointer range. All this is to make a
follow up easier that handles loop-dependent accesses without giving up
on the offset/size completely.