[Attributor] Use dominance early, also when thread effects are not possible
We utilized reachability (expensive) before dominance, now we do it the other
way around. We also allow dominance (and reachability) checks if the pointer
is written for sure, and the read and write happen inside a nosync function.
The reason is that there cannot be any syncronization between the store and
the load, so threading is not an issue here.